Add ScopedQueueSets
ScopedQueueSets contains local and global scoped queue sets.
local scoped queueset holds tasks to be performed on the fetched result (package) that is currently processed and cached locally on the crawler instance.  This avoid refectch and increase the cache hit.
global scoped queueset is the shared queues among crawler instances.
local queueset is popped prior to the global one.  This ensures that cache is utilized before expiration.