Delay fetch result clean up if it is in use
When a fetched result that is cached expires, it may be used as input
by some tools (e.g. scancode) for in-progress processing.  Cleaning
up result at this time causes the tool processing to fail.

Added tracking dependents in fetchResult. When a cached fetchResult
expires, remove it when it is not in use.  If it is still in use, put
it back to cache for another expiration period.

Request has an exisiting clean up mechanism currently used for
cleaning up temporary files or directory when processing of request
is completed.  This can be utilized to remove references to the
underlying cached fetchResult upon processing complete.