Add StorageBackedInMemoryQueue
The graceful shutdown of the crawler is not guaranteed to be executed.
When all crawler loops are busying processing, graceful shutdown is
sometimes not triggered.  In addition, the in-flight requests that are
currently being processed are lost because they are popped and not on
the in memory queue to be published.

Add StorageBackedInMemoryQueue to overcome the above limitations.  It
combines an in memory queue and a storage queue. The queue sequence is
determined by the in memory queue, with the persistence provided by the
storage queue. In the event that graceful shutdown could not be
triggered, the requests on the local queue are already persisted in the
storage queue, and can be picked up by other crawlers when the
visibility time out expires. So no requests are lost.  In the case when
graceful shutdown happens, those requests on the local queue will be
immediately visible to other crawler instances after being published.