Fix parsing LGPL license for some pypi packages
There were issues in parsing LGPL license in spdx-correct. Previously
parsing LGPL was fixed once in "Fixed handling of GNU LGPL licenses in
spdx-correct". The fix depended on the file in /patches directory to
patch spdx-correct.  The /patches directory was not included in the
Dockerfile and hence the previous LGPL fix was not effective in the
docker deployment.

There was a recent release of spdx-correct.  The LGPL issues that the
patch intended to fix seem resolved. Upgrade spdx-correct to the most
recent version. LGPLv2 and LGPLv2+ are still not correctly identified.
Added patch for the specific cases.

Also update Dockerfile so that the patch will be effective in the container
deployment.

Test cases:
        "url": "cd:/pypi/pypi/-/pycountry/22.3.5"
        "url": "cd:/pypi/pypi/-/chardet/5.1.0"
        "url": "cd:/pypi/pypi/-/PyGObject/3.42.0"