Introduce a new traversal policy
The "always" traversal policy behaves as follows:
- if the tool result (e.g. licensee) for a specific component exist, the component will be refetched and the tool will be rerun.
- if the tool result for a specific component is missing, using the "always" policy leads to a "Unreachable for reprocessing" status and the tool being skipped.

The "always" traversal policy is basically a rerun for all the previously ran tools.  It is somewhat cumbersome in the case to retriger harvest, especially for integration tests.

The proposed new policy make reharvest simpler:
- When the tool result for a component is available, the tool will be rerun and tool result updated, similar to the "always" policy.
- When the tool result for a component is not available, the component will be fetched and the tool will be run.
In summary, this "reharvestAlways" policy is to rerun the harvest tools if results exist and run the harvest tools if results are missing.